Changed How Hexa_Stat Values are calculated
Previous iteration was based on averages to achieve a certain value, and the damage gained from achieving said value.

Upon further thought, this would be incorrect to use in a actual optimization because it would be far too dependant on the historical values of the core.

Instead I changed the value to be more of an exact nature where the total boost from leveling up a stat core is used on a more moment to moment basis.

This does mean that this infographic does not consider any rerolling aspect of the hexa-stats.